requirements.txt
alembic==1.15.2
bleach==6.2.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
Flask-Markdown==0.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.2
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
